backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-core==1.34.0
azure-storage-blob==12.19.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.110.0
frozenlist==1.6.2
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
isodate==0.7.2
jiter==0.10.0
multidict==6.4.4
numpy==2.3.0
openai==1.14.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
proglog==0.1.12
propcache==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
realtime==2.4.3
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.27.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==14.2
yarl==1.20.0
azure-identity==1.15.0
passlib==1.7.4
bcrypt==4.1.2
pydantic-settings==2.2.1
